requirements.txt
flask==3.0.0
pandas==2.1.4
openpyxl==3.1.2
google-cloud-storage==2.14.0
google-cloud-bigquery==3.14.1
pyarrow==14.0.2
werkzeug==3.0.1

